Add initial pass at building packages
This includes early versions of systemd unit files for the services
involved in running GitLab, as well as the necessary rake tasks to
actually prepare and build the packages.

Currently, these don't work properly. There's permissions work that
needs to be done and cleanup of the unit files, at least.